schema: int -> tinyint per gender

+++ db/schema.sql
@@ -36,7 +36,7 @@ create table paese (
 );
 
 create table gender (
-	id int unsigned,
+	id tinyint unsigned,
 	nome varchar(50) not null,
 	primary key (id)
 );
@@ -44,7 +44,7 @@ create table gender (
 create table persona (
 	id int unsigned,
 	nome varchar(50) not null,
-	gender int unsigned not null default 0,
+	gender tinyint unsigned not null default 0,
 	immagine varchar(100),
 	data_nascita date,
 	data_morte date,
@@ -126,7 +126,7 @@ create table utente (
 	username varchar(30) unique not null,
 	mail varchar(100) unique,
 	nome varchar(50),
-	gender int unsigned not null default 0,
+	gender tinyint unsigned not null default 0,
 	data_nascita date,
 	password varchar(255) not null,
 	is_admin tinyint unsigned not null default 0,
